Removed constructor arg from Mocker
Had to do this since we can not pass a test case from the static "Data" classes
which are serving as data providers for our test cases.
Instead make Mocker class extend test case class to get access to mockers.
The instance passed to PHPUnit_Framework_MockObject_MockBuilder does not seem
to matter.