Special:UserBitcoinAddresses saves addresses now
The special page uses an UserBitcoinAddressRecordMwDbStore instance to save
user given Bitcoin addresses now.
For testing this better we now inject a MwUserFactory into the
UserBitcoinAddressRecordMwDbStore which is responsible for instantiating User
objects. In production a StandardMwUserFactory is being used, in tests the
UserMocker serves as factory so mocked User objects won't be a problem when
fetching User instances via the store while they actually don't exist in the
database. This would previously result in an User instance being instantiated
by the store on fetching which would not be "equal" to the one used when the
original record has been saved.